One difference in json.h: use "long long" instead of "long" for integer types, or it might be too small to contain Twitter message IDs on for example GCC+i386.

+++ lib/json.h
@@ -77,7 +77,7 @@ typedef struct _json_value
    union
    {
       int boolean;
-      long integer;
+      long long integer;
       double dbl;
 
       struct
